requirements.txt
django-environ
pytest==4.3.0
pytest-django
